[dagit] Split asset graphs onto a new “Plots” tab
