+++ js_modules/dagit/packages/core/src/assets/AssetMaterializationGraphs.tsx
@@ -1,9 +1,14 @@
-import {Box, Colors, NonIdealState, Caption, Subheading, Checkbox} from '@dagster-io/ui';
+import {
+  Box,
+  Colors,
+  NonIdealState,
+  Caption,
+  Subheading,
+  ExternalAnchorButton,
+} from '@dagster-io/ui';
 import flatMap from 'lodash/flatMap';
 import uniq from 'lodash/uniq';
 import * as React from 'react';
-
-import {useStateWithStorage} from '../hooks/useStateWithStorage';
 
 import {AssetValueGraph, AssetValueGraphData} from './AssetValueGraph';
 import {AssetEventGroup} from './groupByPartition';
@@ -22,28 +27,17 @@ export const AssetMaterializationGraphs: React.FC<{
   const graphDataByMetadataLabel = extractNumericData(reversed, props.xAxis);
   const graphLabels = Object.keys(graphDataByMetadataLabel).slice(0, 20).sort();
 
-  const [collapsedLabels, setCollapsedLabels] = useStateWithStorage(
-    'hidden-graphs',
-    validateHiddenGraphsState,
-  );
-
-  const toggleCollapsed = React.useCallback(
-    (label: string) => {
-      setCollapsedLabels((current = []) =>
-        current.includes(label) ? current.filter((c) => c !== label) : [...current, label],
-      );
-    },
-    [setCollapsedLabels],
-  );
+  if (process.env.NODE_ENV === 'test') {
+    return <span />; // chartjs and our useViewport hook don't play nicely with jest
+  }
 
   return (
     <>
       <div
         style={{
-          display: 'flex',
-          flexWrap: 'wrap',
+          display: 'grid',
+          gridTemplateColumns: '1fr 1fr',
           justifyContent: 'stretch',
-          flexDirection: 'column',
         }}
       >
         {graphLabels.map((label) => (
@@ -52,32 +46,23 @@ export const AssetMaterializationGraphs: React.FC<{
             style={{width: '100%'}}
             border={{side: 'bottom', width: 1, color: Colors.KeylineGray}}
           >
-            {props.asSidebarSection ? (
-              <Box padding={{horizontal: 24, top: 8}} flex={{justifyContent: 'space-between'}}>
-                <Caption style={{fontWeight: 700}}>{label}</Caption>
-                <Checkbox
-                  format="switch"
-                  checked={!collapsedLabels.includes(label)}
-                  onChange={() => toggleCollapsed(label)}
-                  size="small"
-                />
-              </Box>
-            ) : (
-              <Box
-                padding={{horizontal: 24, vertical: 16}}
-                border={{side: 'bottom', width: 1, color: Colors.KeylineGray}}
-                flex={{justifyContent: 'space-between'}}
-              >
-                <Subheading>{label}</Subheading>
-                <Checkbox
-                  format="switch"
-                  checked={!collapsedLabels.includes(label)}
-                  onChange={() => toggleCollapsed(label)}
-                  size="small"
-                />
-              </Box>
-            )}
-            {!collapsedLabels.includes(label) ? (
+            <Box
+              style={{width: '100%'}}
+              border={{side: 'right', width: 1, color: Colors.KeylineGray}}
+            >
+              {props.asSidebarSection ? (
+                <Box padding={{horizontal: 24, top: 8}} flex={{justifyContent: 'space-between'}}>
+                  <Caption style={{fontWeight: 700}}>{label}</Caption>
+                </Box>
+              ) : (
+                <Box
+                  padding={{horizontal: 24, vertical: 16}}
+                  border={{side: 'bottom', width: 1, color: Colors.KeylineGray}}
+                  flex={{justifyContent: 'space-between'}}
+                >
+                  <Subheading>{label}</Subheading>
+                </Box>
+              )}
               <Box padding={{horizontal: 24, vertical: 16}}>
                 <AssetValueGraph
                   label={label}
@@ -87,7 +72,7 @@ export const AssetMaterializationGraphs: React.FC<{
                   onHoverX={(x) => x !== xHover && setXHover(x)}
                 />
               </Box>
-            ) : undefined}
+            </Box>
           </Box>
         ))}
       </div>
@@ -103,9 +88,14 @@ export const AssetMaterializationGraphs: React.FC<{
           <Box padding={{horizontal: 24, top: 64}}>
             <NonIdealState
               shrinkable
-              icon="linear_scale"
-              title="No numeric metadata"
-              description={`Include numeric metadata entries in your materializations and observations to see data graphed by ${props.xAxis}.`}
+              icon="asset_plot"
+              title="Asset plots are automatically generated by metadata"
+              description="Include numeric metadata entries in your materializations and observations to see data graphed by time or partition."
+              action={
+                <ExternalAnchorButton href="https://docs.dagster.io/concepts/assets/software-defined-assets#recording-materialization-metadata">
+                  View documentation
+                </ExternalAnchorButton>
+              }
             />
           </Box>
         )
@@ -120,8 +110,6 @@ export const AssetMaterializationGraphs: React.FC<{
     </>
   );
 };
-
-const validateHiddenGraphsState = (json: string[]) => (Array.isArray(json) ? json : []);
 
 /**
  * Helper function that iterates over the asset materializations and assembles time series data
@@ -198,7 +186,6 @@ const extractNumericData = (datapoints: AssetEventGroup[], xAxis: 'time' | 'part
       if (entry.__typename === 'FloatMetadataEntry' && entry.floatValue !== null) {
         y = entry.floatValue;
       }
-
       append(label, {x, y});
     }
   }
